requirements.txt
discord.py==2.4.0
openpyxl==3.1.5
pydantic==2.10.5
SQLAlchemy==2.0.36